large file incoming lol
I proceeded to building a new project website and installed tailwind just to try it and will commit this change. I will try to experiment with tailwind moving forward and there is no turning back.

Let this be a bad example for me in the future and would serve as an example for future devs out there.

P.S. don't do this.